Suggested improvements and bug fixes applied.

+++ Seminars/Sem.07/Pract. 07/Utils.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <ostream>
 
-void myStrCpy(char *, const char *, int = 0);
+void myStrCpy(char *, const char *, size_t = 0);
 int myStrLen(const char *);
 void resizeUtil(char *&, int, bool = false);
 int myStrCmp(const char *, const char *);
